requirements.txt
numpy>=1.20.0
scikit-fuzzy>=0.4.2
matplotlib>=3.5.0
aiogram>=3.0.0
python-dotenv>=0.19.0
asyncpg>=0.27.0
alembic>=1.9.0
sqlalchemy>=2.0.0 